+++ CHANGELOG.md
@@ -8,9 +8,14 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+### Adds
+
+- Adds new `Card` component.
+
 ### Changes
 
 - Exports `VideoPlayerAspectRatios`
+- Renames original `Card` components to be `CardEdition`.
 
 ## 0.23.4